requirements.txt
# Dependências do AO Loot Logger
customtkinter>=5.2.0
scapy>=2.5.0
requests>=2.31.0
pillow>=10.0.0

# Dashboard Web Local
flask>=3.0.0
flask-socketio>=5.3.0

# Para build dos executáveis
pyinstaller>=6.0.0
